+++ app/common/utils.js
@@ -31,6 +31,7 @@ module.exports = {
         return process.env.NODE_ENV === "test";
     },
     timer(eventEmitter, intervalTime) {
+        eventEmitter.setMaxListeners(91);
         setInterval(() => {
             eventEmitter.emit("time-update");
         }, intervalTime);