+++ app/renderer/app_status.js
@@ -1,9 +1,0 @@
-"use strict";
-
-const EventEmitter = require('events');
-
-// TODO: move this to vuex storage
-module.exports = {
-    events: new EventEmitter(),
-    runningTasks: 0
-};