+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+0.5.1 / 2018-06-14
+==================
+
+  * Added contextual menu
+  * Bottom bar on hover style fixed
+  * Improved tabs responsiveness
+  * Fixed bug where more than the max tasks per suite could be added or dragged
+  * Increased timer event emitter max listeners to remove warning
+  * Fixed bug with infinite tasks printing to stdout
+  * Minor styles improvements in tabs
+  * Task card buttons changed
+
 0.5.0 / 2018-06-03
 ==================